+++ internal/component/loki/process/stages/regex_test.go
@@ -33,6 +33,17 @@ stage.regex {
 }
 `
 
+var testRegexAlloyMultiStageWithSourceAndLabelFromGroups = `
+stage.regex {
+    expression = "^(?P<ip>\\S+) (?P<identd>\\S+) (?P<user>\\S+) \\[(?P<timestamp>[\\w:/]+\\s[+\\-]\\d{4})\\] \"(?P<action>\\S+)\\s?(?P<path>\\S+)?\\s?(?P<protocol>\\S+)?\" (?P<status>\\d{3}|-) (?P<size>\\d+|-)\\s?\"?(?P<referer>[^\"]*)\"?\\s?\"?(?P<useragent>[^\"]*)?\"?$"
+	labels_from_groups = true
+}
+stage.regex {
+    expression = "^HTTP\\/(?P<protocol_version>[0-9\\.]+)$"
+    source     = "protocol"
+}	
+`
+
 var testRegexAlloySourceWithMissingKey = `
 stage.json {
     expressions = { "time" = "" }
@@ -64,6 +75,7 @@ func TestPipeline_Regex(t *testing.T) {
 		config          string
 		entry           string
 		expectedExtract map[string]interface{}
+		expectedLables  model.LabelSet
 	}{
 		"successfully run a pipeline with 1 regex stage without source": {
 			testRegexAlloySingleStageWithoutSource,
@@ -81,6 +93,7 @@ func TestPipeline_Regex(t *testing.T) {
 				"referer":   "-",
 				"useragent": "Mozilla/5.0 (Windows; U; Windows NT 5.1; de; rv:1.9.1.7) Gecko/20091221 Firefox/3.5.7 GTB6",
 			},
+			model.LabelSet{},
 		},
 		"successfully run a pipeline with 2 regex stages with source": {
 			testRegexAlloyMultiStageWithSource,
@@ -99,6 +112,38 @@ func TestPipeline_Regex(t *testing.T) {
 				"referer":          "-",
 				"useragent":        "Mozilla/5.0 (Windows; U; Windows NT 5.1; de; rv:1.9.1.7) Gecko/20091221 Firefox/3.5.7 GTB6",
 			},
+			model.LabelSet{},
+		},
+		"successfully run a pipeline with 2 regex stages with source and labels from groups": {
+			testRegexAlloyMultiStageWithSourceAndLabelFromGroups,
+			testRegexLogLine,
+			map[string]interface{}{
+				"ip":               "11.11.11.11",
+				"identd":           "-",
+				"user":             "frank",
+				"timestamp":        "25/Jan/2000:14:00:01 -0500",
+				"action":           "GET",
+				"path":             "/1986.js",
+				"protocol":         "HTTP/1.1",
+				"protocol_version": "1.1",
+				"status":           "200",
+				"size":             "932",
+				"referer":          "-",
+				"useragent":        "Mozilla/5.0 (Windows; U; Windows NT 5.1; de; rv:1.9.1.7) Gecko/20091221 Firefox/3.5.7 GTB6",
+			},
+			model.LabelSet{
+				"ip":        "11.11.11.11",
+				"identd":    "-",
+				"user":      "frank",
+				"timestamp": "25/Jan/2000:14:00:01 -0500",
+				"action":    "GET",
+				"path":      "/1986.js",
+				"protocol":  "HTTP/1.1",
+				"status":    "200",
+				"size":      "932",
+				"referer":   "-",
+				"useragent": "Mozilla/5.0 (Windows; U; Windows NT 5.1; de; rv:1.9.1.7) Gecko/20091221 Firefox/3.5.7 GTB6",
+			},
 		},
 	}
 
@@ -116,6 +161,7 @@ func TestPipeline_Regex(t *testing.T) {
 
 			out := processEntries(pl, newEntry(nil, nil, testData.entry, time.Now()))[0]
 			assert.Equal(t, testData.expectedExtract, out.Extracted)
+			assert.Equal(t, testData.expectedLables, out.Labels)
 		})
 	}
 }
@@ -204,16 +250,55 @@ func TestRegexParser_Parse(t *testing.T) {
 	tests := map[string]struct {
 		config          RegexConfig
 		extracted       map[string]interface{}
+		labels          model.LabelSet
 		entry           string
 		expectedExtract map[string]interface{}
+		expectedLabels  model.LabelSet
 	}{
 		"successfully match expression on entry": {
 			RegexConfig{
 				Expression: "^(?P<ip>\\S+) (?P<identd>\\S+) (?P<user>\\S+) \\[(?P<timestamp>[\\w:/]+\\s[+\\-]\\d{4})\\] \"(?P<action>\\S+)\\s?(?P<path>\\S+)?\\s?(?P<protocol>\\S+)?\" (?P<status>\\d{3}|-) (?P<size>\\d+|-)\\s?\"?(?P<referer>[^\"]*)\"?\\s?\"?(?P<useragent>[^\"]*)?\"?$",
 			},
 			map[string]interface{}{},
+			model.LabelSet{},
 			regexLogFixture,
 			map[string]interface{}{
+				"ip":        "11.11.11.11",
+				"identd":    "-",
+				"user":      "frank",
+				"timestamp": "25/Jan/2000:14:00:01 -0500",
+				"action":    "GET",
+				"path":      "/1986.js",
+				"protocol":  "HTTP/1.1",
+				"status":    "200",
+				"size":      "932",
+				"referer":   "-",
+				"useragent": "Mozilla/5.0 (Windows; U; Windows NT 5.1; de; rv:1.9.1.7) Gecko/20091221 Firefox/3.5.7 GTB6",
+			},
+			model.LabelSet{},
+		},
+		"successfully match expression on entry with label extracted from named capture groups": {
+			RegexConfig{
+				Expression:       "^(?P<ip>\\S+) (?P<identd>\\S+) (?P<user>\\S+) \\[(?P<timestamp>[\\w:/]+\\s[+\\-]\\d{4})\\] \"(?P<action>\\S+)\\s?(?P<path>\\S+)?\\s?(?P<protocol>\\S+)?\" (?P<status>\\d{3}|-) (?P<size>\\d+|-)\\s?\"?(?P<referer>[^\"]*)\"?\\s?\"?(?P<useragent>[^\"]*)?\"?$",
+				LabelsFromGroups: true,
+			},
+			map[string]interface{}{},
+			model.LabelSet{},
+			regexLogFixture,
+			map[string]interface{}{
+				"ip":        "11.11.11.11",
+				"identd":    "-",
+				"user":      "frank",
+				"timestamp": "25/Jan/2000:14:00:01 -0500",
+				"action":    "GET",
+				"path":      "/1986.js",
+				"protocol":  "HTTP/1.1",
+				"status":    "200",
+				"size":      "932",
+				"referer":   "-",
+				"useragent": "Mozilla/5.0 (Windows; U; Windows NT 5.1; de; rv:1.9.1.7) Gecko/20091221 Firefox/3.5.7 GTB6",
+			},
+			model.LabelSet{
 				"ip":        "11.11.11.11",
 				"identd":    "-",
 				"user":      "frank",
@@ -235,8 +320,32 @@ func TestRegexParser_Parse(t *testing.T) {
 			map[string]interface{}{
 				"protocol": "HTTP/1.1",
 			},
+			model.LabelSet{},
 			regexLogFixture,
 			map[string]interface{}{
+				"protocol":         "HTTP/1.1",
+				"protocol_version": "1.1",
+			},
+			model.LabelSet{},
+		},
+		"successfully match expression on extracted[source] with label extracted from named capture groups": {
+			RegexConfig{
+				Expression:       "^HTTP\\/(?P<protocol_version>.*)$",
+				Source:           &protocolStr,
+				LabelsFromGroups: true,
+			},
+			map[string]interface{}{
+				"protocol": "HTTP/1.1",
+			},
+			model.LabelSet{
+				"protocol": "HTTP/1.1",
+			},
+			regexLogFixture,
+			map[string]interface{}{
+				"protocol":         "HTTP/1.1",
+				"protocol_version": "1.1",
+			},
+			model.LabelSet{
 				"protocol":         "HTTP/1.1",
 				"protocol_version": "1.1",
 			},
@@ -246,8 +355,10 @@ func TestRegexParser_Parse(t *testing.T) {
 				Expression: "^(?s)(?P<time>\\S+?) (?P<stream>stdout|stderr) (?P<flags>\\S+?) (?P<message>.*)$",
 			},
 			map[string]interface{}{},
+			model.LabelSet{},
 			"blahblahblah",
 			map[string]interface{}{},
+			model.LabelSet{},
 		},
 		"failed to match expression on extracted[source]": {
 			RegexConfig{
@@ -257,20 +368,24 @@ func TestRegexParser_Parse(t *testing.T) {
 			map[string]interface{}{
 				"protocol": "unknown",
 			},
+			model.LabelSet{},
 			"unknown/unknown",
 			map[string]interface{}{
 				"protocol": "unknown",
 			},
+			model.LabelSet{},
 		},
 		"case insensitive": {
 			RegexConfig{
 				Expression: "(?i)(?P<bad>panic:|core_dumped|failure|error|attack| bad |illegal |denied|refused|unauthorized|fatal|failed|Segmentation Fault|Corrupted)",
 			},
 			map[string]interface{}{},
+			model.LabelSet{},
 			"A Terrible Error has occurred!!!",
 			map[string]interface{}{
 				"bad": "Error",
 			},
+			model.LabelSet{},
 		},
 		"missing extracted[source]": {
 			RegexConfig{
@@ -278,8 +393,10 @@ func TestRegexParser_Parse(t *testing.T) {
 				Source:     &protocolStr,
 			},
 			map[string]interface{}{},
+			model.LabelSet{},
 			"blahblahblah",
 			map[string]interface{}{},
+			model.LabelSet{},
 		},
 		"invalid data type in extracted[source]": {
 			RegexConfig{
@@ -289,10 +406,12 @@ func TestRegexParser_Parse(t *testing.T) {
 			map[string]interface{}{
 				"protocol": true,
 			},
+			model.LabelSet{},
 			"unknown/unknown",
 			map[string]interface{}{
 				"protocol": true,
 			},
+			model.LabelSet{},
 		},
 	}
 	for tName, tt := range tests {
@@ -304,8 +423,9 @@ func TestRegexParser_Parse(t *testing.T) {
 			if err != nil {
 				t.Fatalf("failed to create regex parser: %s", err)
 			}
-			out := processEntries(p, newEntry(tt.extracted, nil, tt.entry, time.Now()))[0]
+			out := processEntries(p, newEntry(tt.extracted, tt.labels, tt.entry, time.Now()))[0]
 			assert.Equal(t, tt.expectedExtract, out.Extracted)
+			assert.Equal(t, tt.expectedLabels, out.Labels)
 		})
 	}
 }